step by step 5 back 2 forward

+++ src/VoxelBlock.cpp
@@ -19,9 +19,7 @@ VoxelBlock::VoxelBlock(uint32 lod, iaVector3I position, iaVector3I parentAdress)
     _parentAdress = parentAdress;
 
     float64 halfSize = static_cast<float64>(_size >> 1);
-    _blockCenterPos.set(static_cast<float64>(_position._x) + halfSize,
-        static_cast<float64>(_position._y) + halfSize,
-        static_cast<float64>(_position._z) + halfSize);
+    _blockCenterPos.set(_position._x + halfSize, _position._y + halfSize, _position._z + halfSize);
     
     for (int i = 0; i < 8; ++i)
 	{
